requirements.txt
colorama==0.4.4
diff_match_patch==20230430
diskcache==5.6.3
fastapi==0.110.1
openai==1.17.1
pydantic==2.7.0
python-dotenv==1.0.1
websockets==11.0.3
uvicorn==0.22.0